+++ NEWS.md
@@ -1,3 +1,9 @@
+# spant 3.8.0
+* Added deface option to fit_svs.
+* Added overwrite option to fit_svs.
+* fit_svs and fit_svs_edited now accept vectorised file inputs for batch
+analyses.
+
 # spant 3.7.0
 * Added lb_renoise function for fMRS analyses.
 * Added paths2df function to help extract ID fields from a vector of paths.